added Persistence test added user fk in Meal and changed Id to long

+++ src/main/java/com/nutrition/sweng/Inbound/MealController.java
@@ -20,7 +20,7 @@ public class MealController {
     }
 
     @GetMapping("/{id}")
-    public MealDto getMeal(@PathVariable Integer id){
+    public MealDto getMeal(@PathVariable Long id){
         Meal meal = this.mealService.getMeal(id);
         return new MealDto(meal);
     }
